additions from the gsta script

+++ DBSCAN/run_dbscan_postgres_by_uid.py
@@ -15,70 +15,18 @@ from sqlalchemy import create_engine
 
 import random
 
-#%%
-import aws_credentials as a_c
-user = a_c.user
-host = a_c.host
-port = '5432'
-database = 'aws_ais_clustering'
-password = a_c.password
+# Geo-Spatial Temporal Analysis package
+import gsta
+import gsta_config
 
-aws_conn = psycopg2.connect(host=host,database=database, user=user,password=password)
-aws_c = aws_conn.cursor()
-if aws_c:
-    print('Connection to AWS is good.'.format(database))
-else: print('Connection failed.')
-aws_c.close()
+aws_conn = gsta.connect_psycopg2(gsta_config.aws_ais_cluster_params)
+loc_conn = gsta.connect_psycopg2(gsta_config.loc_cargo_params)
+aws_conn.close()    
 
-
-# def create_aws_engine(database):
-#     import aws_credentials as a_c
-#     user = a_c.user
-#     host = a_c.host
-#     port = '5432'
-#     password = a_c.password
-#     try:
-#         aws_engine = create_engine('postgresql://{}:{}@{}:{}/{}'.format(user, password, host, port, database))
-#         print('AWS Engine created and connected.')
-#         return aws_engine
-#     except:
-#         print('AWS Engine creation failed.')
-#         return None
-
-# aws_engine = create_aws_engine('aws_ais_clustering')
-
-#%%
-database='ais_test'
-loc_conn = psycopg2.connect(host="localhost",database=database)
-c = loc_conn.cursor()
-if c:
-    print('Connection to {} is good.'.format(database))
-else:
-    print('Error connecting.')
-c.close()
-
-
-#%% Get mmsis to cluster
-
-# all mmsis
-source_table = 'ship_trips'
-c = loc_conn.cursor()
-c.execute("""SELECT DISTINCT(mmsi) FROM {};""".format(source_table))
-mmsi_list = c.fetchall()
-c.close()
-#%%
-# all tankers and cargo ships
-c = loc_conn.cursor()
-c.execute("""SELECT DISTINCT(mmsi)
-          FROM cargo_tanker_mmsis;""")
-cargo_tanker_mmsi_list = c.fetchall()
-c.close()
-
-cargo_tanker_mmsi_sample = random.sample(cargo_tanker_mmsi_list,200)
 
 
 #%%
-def postgres_dbscan(source_table, new_table_name, eps_km, min_samples, 
+def postgres_dbscan(source_table, new_table_name, eps, min_samples, 
                     mmsi_list, conn, lat='lat', lon='lon'):
     
     c = conn.cursor()
@@ -92,13 +40,13 @@ def postgres_dbscan(source_table, new_table_name, eps_km, min_samples,
     );""".format(new_table_name))
     conn.commit()
 
-    print("""Starting processing on DBSCAN with eps_km={} and
-          min_samples={} """.format(str(eps_km), str(min_samples)))
+    print("""Starting processing on DBSCAN with eps={} and
+          min_samples={} """.format(str(eps), str(min_samples)))
     
     for mmsi in mmsi_list:          
         dbscan_sql = """INSERT INTO {0} (id, mmsi, lat, lon, clust_id)
         SELECT id, mmsi, {1}, {2},
-        ST_ClusterDBSCAN(Geometry(geog), eps := {3}, minpoints := {4})
+        ST_ClusterDBSCAN(geom, eps := {3}, minpoints := {4})
         over () as clust_id
         FROM {5}
         WHERE mmsi = '{6}';""".format(new_table_name, lat, lon, str(eps), str(min_samples), source_table, mmsi[0])
@@ -118,46 +66,47 @@ def make_tables_geom(table, conn):
     # add a geom column to the new table and populate it from the lat and lon columns
     c = conn.cursor()
     c.execute("""ALTER TABLE {} ADD COLUMN
-                geog geography(Point, 4326);""".format(table))
+                geom geometry(Point, 4326);""".format(table))
     conn.commit()
     c.execute("""UPDATE {} SET
-                geog = ST_SetSRID(ST_MakePoint(lon, lat), 4326);""".format(table))
+                geom = ST_SetSRID(ST_MakePoint(lon, lat), 4326);""".format(table))
     conn.commit()
     c.close()
 
 
 #%%
+drop = False
+source_table = 'cargo_ship_position'
+
+c = loc_conn.cursor()
+c.execute("""SELECT DISTINCT(mmsi) FROM {};""".format(source_table))
+mmsi_list = c.fetchall()
+c.close()
 
 print('Function run at:', datetime.datetime.now())
-epsilons = [1, 2, 5, 7]
+epsilons = [500, 1000, 2000, 5000]
 samples = [10, 25, 50, 100, 250, 500]
 
-for eps_km in epsilons:
-    for min_samples in samples:
-
-        loc_conn = psycopg2.connect(host="localhost",database=database)
+for eps in epsilons:
+    for min_samples in samples:        
         tick = datetime.datetime.now()
-        
-        #this formulation will yield epsilon based on km desired
-        kms_per_radian = 6371.0088
-        eps = eps_km / kms_per_radian
-    
         # make the new table name
-        new_table_name = ('dbscan_results_cargo_by_mmsi_' + str(eps_km).replace('.','_') +
+        new_table_name = ('dbscan_results_by_mmsi_' + str(eps).replace('.','_') +
                           '_' + str(min_samples))
-        
-        # drop table if exists if needed.
-        c = loc_conn.cursor()
-        c.execute("""DROP TABLE IF EXISTS {}""".format(new_table_name))
-        loc_conn.commit()
-        c.close()
-        
-        # pass the epsilon in km.  the function will convert it to radians
-        postgres_dbscan('cargo_tanker_position', new_table_name, eps_km , min_samples, 
-                        cargo_tanker_mmsi_sample, loc_conn)
+        if drop == True:
+            # drop table if exists if needed.
+            c = loc_conn.cursor()
+            c.execute("""DROP TABLE IF EXISTS {}""".format(new_table_name))
+            loc_conn.commit()
+            c.close()
+
+        postgres_dbscan('cargo_tanker_position', new_table_name, eps , min_samples, 
+                        mmsi_list, loc_conn)
         
         # add geom colum to the new tables
         make_tables_geom(new_table_name, loc_conn)
+        
+        loc_conn.close()
 
         #timekeeping
         tock = datetime.datetime.now()
